refactor: extract PickSlip component from order detail page
Pure refactor: no visual or behavioural change on the detail page.
PickSlip is a presentational component that takes a JSON-shaped
order, tenant, and lines, plus an optional refundsSlot for the
detail page's refunds block. Sets up Task 2 (batch print).
